refactor(domain): domain.card로 위치되는 클래스 이동
