+++ src/test/java/domain/CardMachineTest.java
@@ -3,6 +3,8 @@ package domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import domain.card.Card;
+import domain.card.CardMachine;
 import java.util.HashMap;
 import java.util.Map;
 import org.junit.jupiter.api.Nested;
